python_requirement_packages.txt
colorama==0.4.6
DateTime==5.3
fuzzywuzzy==0.18.0
greenlet==3.0.2
mysql-connector==2.2.9
mysql-connector-python==8.2.0
numpy==1.26.2
pandas==2.1.4
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3.post1
rapidfuzz==3.5.2
six==1.16.0
SQLAlchemy==2.0.23
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
zope.interface==6.1
